More changes to file system
iOS: Due to the way files are accessed in iOS, every time you click save you give the name of the file. Also, you cannot save to overwrite an external file so if you load a file not in the cews folder then you will have to save it as another file in the cews folder.

Android: If you load a file or save a new file then that path is then saved meaning that any time you press the save button it will save to that path. The consequence of this is that if you want to start a new file after having loaded/saved beforehand, then you must restart the app.